add invariant in class Time, and changed typeof field seconds on long int

+++ src/libcsc/libcsc/time/time.cpp
@@ -25,11 +25,11 @@ TimeSpan operator-(const Time &lhs, const Time &rhs) {
 }
 
 std::ostream &operator<<(std::ostream &os, const Time &t) {
-  const size_t all_secs = t.get_time();
-  const size_t days = all_secs / static_cast<size_t>((24 * 3600));
-  const size_t hours = (all_secs / 3600) - (days * 24);
-  const size_t minutes = (all_secs % 3600) / 60;
-  const size_t seconds = all_secs % 60;
+  const long int all_secs = t.get_time();
+  const long int days = all_secs / static_cast<long>(24 * 3600);
+  const long int hours = (all_secs / 3600) - (days * 24);
+  const long int minutes = (all_secs % 3600) / 60;
+  const long int seconds = all_secs % 60;
 
   return os << days << "d " << hours << "h " << minutes << "m " << seconds
             << "s\n";
@@ -77,10 +77,18 @@ bool operator>=(const TimeSpan &lhs, const TimeSpan &rhs) {
 
 namespace literals {
 
-Time operator"" _d(unsigned long long days) { return {days * 24 * 3600}; }
-Time operator"" _h(unsigned long long hours) { return {hours * 3600}; }
-Time operator"" _m(unsigned long long minutes) { return {minutes * 60}; }
-Time operator"" _s(unsigned long long seconds) { return {seconds}; }
+Time operator"" _d(unsigned long long days) {
+  return {static_cast<long>(days * 24 * 3600)};
+}
+Time operator"" _h(unsigned long long hours) {
+  return {static_cast<long>(hours * 3600)};
+}
+Time operator"" _m(unsigned long long minutes) {
+  return {static_cast<long>(minutes * 60)};
+}
+Time operator"" _s(unsigned long long seconds) {
+  return {static_cast<long>(seconds)};
+}
 
 } // namespace literals
